+++ lib-limeade/src/main/java/uk/gov/account/ipv/cri/lime/limeade/service/http/mtls/KeyCertHelper.java
@@ -1,0 +1,55 @@
+package uk.gov.account.ipv.cri.lime.limeade.service.http.mtls;
+
+import uk.gov.account.ipv.cri.lime.limeade.annotation.ExcludeConstructorFromGeneratedCoverageReport;
+
+import java.io.ByteArrayInputStream;
+import java.security.KeyFactory;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+import java.security.PrivateKey;
+import java.security.cert.CertificateEncodingException;
+import java.security.cert.CertificateException;
+import java.security.cert.CertificateFactory;
+import java.security.cert.X509Certificate;
+import java.security.spec.InvalidKeySpecException;
+import java.security.spec.PKCS8EncodedKeySpec;
+import java.util.Base64;
+
+public class KeyCertHelper {
+
+    @ExcludeConstructorFromGeneratedCoverageReport
+    private KeyCertHelper() {
+        throw new IllegalStateException("Instantiation is not valid for this class.");
+    }
+
+    public static X509Certificate getDecodedX509Certificate(String base64String)
+            throws CertificateException {
+        byte[] binaryCertificate = Base64.getDecoder().decode(base64String);
+        CertificateFactory factory = CertificateFactory.getInstance("X.509");
+        return (X509Certificate)
+                factory.generateCertificate(new ByteArrayInputStream(binaryCertificate));
+    }
+
+    public static PrivateKey getDecodedPrivateRSAKey(String base64String)
+            throws NoSuchAlgorithmException, InvalidKeySpecException {
+        byte[] binaryKey = Base64.getDecoder().decode(base64String);
+        KeyFactory factory = KeyFactory.getInstance("RSA");
+        PKCS8EncodedKeySpec privateKeySpec = new PKCS8EncodedKeySpec(binaryKey);
+        return factory.generatePrivate(privateKeySpec);
+    }
+
+    public static Thumbprints makeThumbprint(X509Certificate certificate)
+            throws CertificateEncodingException, NoSuchAlgorithmException {
+        return new Thumbprints(
+                getCertThumbprint(certificate, "SHA-1"), getCertThumbprint(certificate, "SHA-256"));
+    }
+
+    private static String getCertThumbprint(X509Certificate cert, String hashAlg)
+            throws NoSuchAlgorithmException, CertificateEncodingException {
+        MessageDigest md = MessageDigest.getInstance(hashAlg);
+        byte[] der = cert.getEncoded();
+        md.update(der);
+        byte[] digest = md.digest();
+        return Base64.getUrlEncoder().withoutPadding().encodeToString(digest);
+    }
+}